Refactoring agents.py
