+++ src/smolagents/agents.py
@@ -18,7 +18,6 @@ import importlib
 import inspect
 import json
 import os
-import re
 import tempfile
 import textwrap
 import time
@@ -30,7 +29,6 @@ from typing import Any, Callable, Dict, Generator, List, Optional, Set, Tuple, T
 import jinja2
 import yaml
 from huggingface_hub import create_repo, metadata_update, snapshot_download, upload_folder
-from jinja2 import StrictUndefined, Template
 from rich.console import Group
 from rich.panel import Panel
 from rich.rule import Rule
@@ -64,101 +62,9 @@ from .utils import (
     parse_json_tool_call,
     truncate_content,
 )
-
+from .templates import EMPTY_PROMPT_TEMPLATES, APP_TEMPLATE, PromptTemplates, populate_template
 
 logger = getLogger(__name__)
-
-
-def get_variable_names(self, template: str) -> Set[str]:
-    pattern = re.compile(r"\{\{([^{}]+)\}\}")
-    return {match.group(1).strip() for match in pattern.finditer(template)}
-
-
-def populate_template(template: str, variables: Dict[str, Any]) -> str:
-    compiled_template = Template(template, undefined=StrictUndefined)
-    try:
-        return compiled_template.render(**variables)
-    except Exception as e:
-        raise Exception(f"Error during jinja template rendering: {type(e).__name__}: {e}")
-
-
-class PlanningPromptTemplate(TypedDict):
-    """
-    Prompt templates for the planning step.
-
-    Args:
-        initial_facts (`str`): Initial facts prompt.
-        initial_plan (`str`): Initial plan prompt.
-        update_facts_pre_messages (`str`): Update facts pre-messages prompt.
-        update_facts_post_messages (`str`): Update facts post-messages prompt.
-        update_plan_pre_messages (`str`): Update plan pre-messages prompt.
-        update_plan_post_messages (`str`): Update plan post-messages prompt.
-    """
-
-    initial_facts: str
-    initial_plan: str
-    update_facts_pre_messages: str
-    update_facts_post_messages: str
-    update_plan_pre_messages: str
-    update_plan_post_messages: str
-
-
-class ManagedAgentPromptTemplate(TypedDict):
-    """
-    Prompt templates for the managed agent.
-
-    Args:
-        task (`str`): Task prompt.
-        report (`str`): Report prompt.
-    """
-
-    task: str
-    report: str
-
-
-class FinalAnswerPromptTemplate(TypedDict):
-    """
-    Prompt templates for the final answer.
-
-    Args:
-        pre_messages (`str`): Pre-messages prompt.
-        post_messages (`str`): Post-messages prompt.
-    """
-
-    pre_messages: str
-    post_messages: str
-
-
-class PromptTemplates(TypedDict):
-    """
-    Prompt templates for the agent.
-
-    Args:
-        system_prompt (`str`): System prompt.
-        planning ([`~agents.PlanningPromptTemplate`]): Planning prompt templates.
-        managed_agent ([`~agents.ManagedAgentPromptTemplate`]): Managed agent prompt templates.
-        final_answer ([`~agents.FinalAnswerPromptTemplate`]): Final answer prompt templates.
-    """
-
-    system_prompt: str
-    planning: PlanningPromptTemplate
-    managed_agent: ManagedAgentPromptTemplate
-    final_answer: FinalAnswerPromptTemplate
-
-
-EMPTY_PROMPT_TEMPLATES = PromptTemplates(
-    system_prompt="",
-    planning=PlanningPromptTemplate(
-        initial_facts="",
-        initial_plan="",
-        update_facts_pre_messages="",
-        update_facts_post_messages="",
-        update_plan_pre_messages="",
-        update_plan_post_messages="",
-    ),
-    managed_agent=ManagedAgentPromptTemplate(task="", report=""),
-    final_answer=FinalAnswerPromptTemplate(pre_messages="", post_messages=""),
-)
 
 
 class MultiStepAgent:
@@ -710,44 +616,7 @@ You have been provided with these additional arguments, that you can access usin
         # Make agent.py file with Gradio UI
         agent_name = f"agent_{self.name}" if getattr(self, "name", None) else "agent"
         managed_agent_relative_path = relative_path + "." if relative_path is not None else ""
-        app_template = textwrap.dedent("""
-            import yaml
-            import os
-            from smolagents import GradioUI, {{ class_name }}, {{ agent_dict['model']['class'] }}
-
-            # Get current directory path
-            CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
-
-            {% for tool in tools.values() -%}
-            from {{managed_agent_relative_path}}tools.{{ tool.name }} import {{ tool.__class__.__name__ }} as {{ tool.name | camelcase }}
-            {% endfor %}
-            {% for managed_agent in managed_agents.values() -%}
-            from {{managed_agent_relative_path}}managed_agents.{{ managed_agent.name }}.app import agent_{{ managed_agent.name }}
-            {% endfor %}
-
-            model = {{ agent_dict['model']['class'] }}(
-            {% for key in agent_dict['model']['data'] if key not in ['class', 'last_input_token_count', 'last_output_token_count'] -%}
-                {{ key }}={{ agent_dict['model']['data'][key]|repr }},
-            {% endfor %})
-
-            {% for tool in tools.values() -%}
-            {{ tool.name }} = {{ tool.name | camelcase }}()
-            {% endfor %}
-
-            with open(os.path.join(CURRENT_DIR, "prompts.yaml"), 'r') as stream:
-                prompt_templates = yaml.safe_load(stream)
-
-            {{ agent_name }} = {{ class_name }}(
-                model=model,
-                tools=[{% for tool_name in tools.keys() if tool_name != "final_answer" %}{{ tool_name }}{% if not loop.last %}, {% endif %}{% endfor %}],
-                managed_agents=[{% for subagent_name in managed_agents.keys() %}agent_{{ subagent_name }}{% if not loop.last %}, {% endif %}{% endfor %}],
-                {% for attribute_name, value in agent_dict.items() if attribute_name not in ["model", "tools", "prompt_templates", "authorized_imports", "managed_agents", "requirements"] -%}
-                {{ attribute_name }}={{ value|repr }},
-                {% endfor %}prompt_templates=prompt_templates
-            )
-            if __name__ == "__main__":
-                GradioUI({{ agent_name }}).launch()
-            """).strip()
+        app_template = textwrap.dedent(APP_TEMPLATE).strip()
         template_env = jinja2.Environment(loader=jinja2.BaseLoader(), undefined=jinja2.StrictUndefined)
         template_env.filters["repr"] = repr
         template_env.filters["camelcase"] = lambda value: "".join(word.capitalize() for word in value.split("_"))
